공통 로직 hooks로 분리
