+++ src/pages/PollDetail/PollResultPage.tsx
@@ -1,9 +1,8 @@
 import { useNavigate, useParams } from 'react-router-dom';
-import { useGetPost } from '@/api/useGetPost';
-import { useGetResult } from '@/api/useGetResult';
 import { Header } from '@/components/common/Header/Header';
 import Icon from '@/components/common/Icon';
 import ResultItem from '@/components/poll-detail/Result/ResultItem';
+import { usePollResult } from '@/components/poll-detail/hooks';
 
 export default function PollResultPage() {
   const navigate = useNavigate();
@@ -13,8 +12,16 @@ export default function PollResultPage() {
     return <div>없는 게시글이용</div>;
   }
 
-  const { data: post } = useGetPost(postId);
-  const { data: result } = useGetResult(postId);
+  const { post, sortedChoices, totalVotes, calculatePercentage, isLoading } =
+    usePollResult(postId);
+
+  if (isLoading) {
+    return (
+      <div className="w-full h-screen flex items-center justify-center text-gray-600">
+        로딩 중이용
+      </div>
+    );
+  }
 
   if (!post) {
     return (
@@ -23,23 +30,6 @@ export default function PollResultPage() {
       </div>
     );
   }
-
-  const resultChoices = (result ?? []).map((result) => ({
-    id: result.id,
-    title: result.title,
-    imageUrl: result.imageUrl,
-    voteCount: result.voteCount,
-    voteRatio: result.voteRatio,
-  }));
-
-  // 투표 수로 정렬 (내림차순)
-  const sortedChoices = [...resultChoices].sort(
-    (a, b) => b.voteCount - a.voteCount,
-  );
-  const totalVotes = sortedChoices.reduce(
-    (sum, choice) => sum + choice.voteCount,
-    0,
-  );
 
   return (
     <div className="w-full bg-gray-100 min-h-screen flex flex-col pt-[80px] relative">
@@ -68,8 +58,7 @@ export default function PollResultPage() {
             <span className="text-body-2">{totalVotes}표</span>
           </div>
         </div>
-
-        {resultChoices.length > 0 && (
+        {sortedChoices.length > 0 && (
           <div className="space-y-4">
             {sortedChoices.map((choice, index) => (
               <ResultItem
@@ -78,9 +67,7 @@ export default function PollResultPage() {
                 title={choice.title}
                 imageUrl={choice.imageUrl}
                 voteCount={choice.voteCount}
-                percentage={
-                  totalVotes > 0 ? (choice.voteCount / totalVotes) * 100 : 0
-                }
+                percentage={calculatePercentage(choice.voteCount)}
               />
             ))}
           </div>
